feat: windows installer using docker compose (#1297)
* install docker desktop on windows

* fix windows installation

* remove empty file

* move command related functions to cmd.go

* fix windows installer

* use build tags instead of isWindows function

* mute refreshenv and fix windows conditional

* remove +build tags
